requirements.txt
openai
langchain>=0.0.290
pandas
sqlalchemy
python-dotenv
requests
tabulate
serpapi